\name{FDAdesign}
\alias{FDAdesign}
\alias{FDAdesign}
\docType{package}
\title{FDAdesign}
\description{
Optimal Sampling Design for Functional Data
}
\details{
\tabular{ll}{
Package: \tab FDAdesign\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-09-09\cr
License: GPL-2
}
}
\author{
So Young Park \email{spark13@ncsu.edu},
Luo Xiao \email{lxiao5@ncsu.edu},
Ana-Maria Staicu \email{astaicu@ncsu.edu}

}
\keyword{ package }

