% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plot.R
\name{interactive_plot}
\alias{interactive_plot}
\title{Optimal Sampling Design for Functional Data Analysis - interactive plots}
\usage{
interactive_plot(result, labels_candidate_pts = NULL)
}
\arguments{
\item{result}{list of results of \code{opt_design_fda} for p = 1, 2, 3}

\item{labels_candidate_pts}{candidate points for x-label on interactive plots; 
default = NULL is equivalent to 1:number of candidate points.}
}
\value{
user interface illustrating objective functions for p = 1, 2, 3.
}
\description{
Produces interactive graphics to visualize objective functions for selecting
optimal sampling points for p = 1, 2, and 3.
}
\examples{
# see example for opt_design_fda() function
}
\author{
So Young Park \email{spark13@ncsu.edu},
Luo Xiao \email{lxiao5@ncsu.edu},
Ana-Maria Staicu \email{astaicu@ncsu.edu}
}
\seealso{
\code{\link{opt_design_fda}} /
\code{\link{selection_p}} /
\code{\link{interactive_plot}}
}

