% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_p.R
\name{selection_p}
\alias{selection_p}
\title{Optimal Sampling Design for Functional Data Analysis - selection of number of optimal points}
\usage{
selection_p(delta = 0.1, pmax = 10, Phi, lambda, B = NULL,
  sigma2 = 10^-14)
}
\arguments{
\item{delta}{selection parameter}

\item{Phi}{d by L matrix of eigenfunctions evaluated at d candidate points; L is number of PCs.}

\item{lambda}{eigenvalues; a vector of length L}

\item{B}{design criterion matrix (e.g. for recovering curves, B = diag(L); a square matrix with dim = L}

\item{sigma2}{measurement error variance associated with functional object.}
}
\value{
plot  plot of error levels corresponding to \code{p_vec}.

p.sel  number of optimal points, p, determined with given threshold.

opt.sel  result of \link{opt_design_fda} corresponding to the selected p, \code{p.sel}.

opt_result results of \link{opt_design_fda} corresponding to \code{p_vec}.

INPUT input of \code{selection_p} provided as input.
}
\description{
Selects number of optimal points, p. 
This function also returns results of the \link{opt_design_fda} function for {p = 1, ..., p.sel+1}.
}
\examples{
# see example for opt_design_fda function
}
\author{
So Young Park \email{spark13@ncsu.edu},
Luo Xiao \email{lxiao5@ncsu.edu},
Ana-Maria Staicu \email{astaicu@ncsu.edu}
}
\seealso{
\code{\link{opt_design_fda}} / \code{\link{selection_p}} / \code{\link{interactive_plot}}
}

