% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection_p.R
\name{selection_p}
\alias{selection_p}
\title{Optimal Sampling Design for Functional Data Analysis - selection of number of optimal points}
\usage{
selection_p(p_vec = 1:5, threshold = 5, Phi, lambda, B = NULL,
  sigma2 = 10^-14)
}
\arguments{
\item{p_vec}{a vector of p's}

\item{threshold}{user-specified threshold h for selecting p (relative error levels)}

\item{Phi}{d by L matrix of eigenfunctions evaluated at d candidate points; L is number of PCs.}

\item{lambda}{eigenvalues; a vector of length L}

\item{B}{design criterion matrix (e.g. for recovering curves, B = diag(L); a square matrix with dim = L}

\item{sigma2}{measurement error variance associated with functional object.}
}
\value{
plot  plot of error levels corresponding to \code{p_vec}.

p.sel  number of optimal points, p, determined with given threshold.

opt.sel  result of \link{opt_design_fda} corresponding to the selected p, \code{p.sel}.

opt_result results of \link{opt_design_fda} corresponding to \code{p_vec}.

INPUT input of \code{selection_p} provided as input.
}
\description{
Selects number of optimal points, p, from a vector of p's (\code{p_vec}) given as input.
This function also returns results of the \link{opt_design_fda} function for given p's (\code{p_vec}).
}
\examples{
# see example for opt_design_fda function
}
\author{
So Young Park \email{spark13@ncsu.edu},
Luo Xiao \email{lxiao5@ncsu.edu},
Ana-Maria Staicu \email{astaicu@ncsu.edu}
}
\references{
\url{}
}
\seealso{
\code{\link{opt_design_fda}} / \code{\link{selection_p}} / \code{\link{interactive_plot}}
}

